Streamlabs: Split Donation/Alert/User into diffrent Endpoints
