+++ src/main/java/me/philippheuer/twitch4j/streamlabs/endpoints/AlertEndpoint.java
@@ -1,0 +1,77 @@
+package me.philippheuer.twitch4j.streamlabs.endpoints;
+
+import com.jcabi.log.Logger;
+import lombok.Getter;
+import lombok.Setter;
+import me.philippheuer.twitch4j.auth.model.OAuthCredential;
+import me.philippheuer.twitch4j.helper.QueryRequestInterceptor;
+import me.philippheuer.twitch4j.streamlabs.StreamlabsClient;
+import me.philippheuer.twitch4j.streamlabs.model.AlertCreate;
+import org.springframework.web.client.RestTemplate;
+
+import java.util.Arrays;
+import java.util.Optional;
+
+@Getter
+@Setter
+public class AlertEndpoint extends AbstractStreamlabsEndpoint {
+
+	/**
+	 * Holds the credentials to the current user
+	 */
+	private OAuthCredential oAuthCredential;
+
+	/**
+	 * Stream Labs - Authenticated Endpoint
+	 */
+	public AlertEndpoint(StreamlabsClient streamlabsClient, OAuthCredential credential) {
+		super(streamlabsClient);
+		setOAuthCredential(credential);
+	}
+
+	/**
+	 * Endpoint: Create Alert
+	 * Trigger a custom alert for the authenticated user.
+	 * Requires Scope: alerts.create
+	 *
+	 * @param type               This parameter determines which alert box this alert will show up in, and thus should be one of the following: follow, subscription, donation, or host
+	 * @param message            The message to show with this alert. If not supplied, no message will be shown. Surround special tokens with *s, for example: This is my *special* alert!
+	 * @param duration           How many seconds this alert should be displayed.
+	 * @param special_text_color The color to use for special tokens. Must be a valid CSS color string.
+	 * @param imageUrl           The href pointing to an image resource to play when this alert shows. If an empty string is supplied, no image will be displayed.
+	 * @param soundUrl           The href pointing to a sound resource to play when this alert shows. If an empty string is supplied, no sound will be played.
+	 * @return Success?
+	 */
+	public Boolean createAlert(String type, Optional<String> message, Optional<Integer> duration, Optional<String> special_text_color, Optional<String> imageUrl, Optional<String> soundUrl) {
+		// Validate Parameters
+		if (!Arrays.asList("follow", "subscription", "donation", "host").contains(type)) {
+			throw new RuntimeException("Invalid Type");
+		}
+
+		// Endpoint
+		String requestUrl = String.format("%s/alerts", getStreamlabsClient().getEndpointUrl());
+		RestTemplate restTemplate = getStreamlabsClient().getRestClient().getRestTemplate();
+
+		// Parameters
+		restTemplate.getInterceptors().add(new QueryRequestInterceptor("access_token", getOAuthCredential().getOAuthToken()));
+		restTemplate.getInterceptors().add(new QueryRequestInterceptor("type", type));
+		restTemplate.getInterceptors().add(new QueryRequestInterceptor("message", message.orElse("")));
+		restTemplate.getInterceptors().add(new QueryRequestInterceptor("duration", duration.orElse(10).toString()));
+		restTemplate.getInterceptors().add(new QueryRequestInterceptor("special_text_color", special_text_color.orElse("")));
+		restTemplate.getInterceptors().add(new QueryRequestInterceptor("image_href", imageUrl.orElse("")));
+		restTemplate.getInterceptors().add(new QueryRequestInterceptor("sound_href", soundUrl.orElse("")));
+
+		// REST Request
+		try {
+			AlertCreate responseObject = restTemplate.getForObject(requestUrl, AlertCreate.class);
+
+			Logger.debug(this, "Sreamlabs: Created new Alert for %s", getOAuthCredential().getDisplayName());
+
+			return responseObject.getSuccess();
+		} catch (Exception ex) {
+			ex.printStackTrace();
+			return null;
+		}
+	}
+
+}
