+++ src/main/resources/logback.xml
@@ -2,12 +2,12 @@
 
 <configuration scan="true">
 	<!-- Avoid logger Spam -->
-	<logger name="org.eclipse.jetty" level="INFO" />
-	<logger name="io.netty" level="INFO" />
-	<logger name="org.springframework.web.client.RestTemplate" level="INFO" />
+	<logger name="org.eclipse.jetty" level="INFO"/>
+	<logger name="io.netty" level="INFO"/>
+	<logger name="org.springframework.web.client.RestTemplate" level="INFO"/>
 
 	<!-- Configure Twitch4J logger level -->
-	<logger name="me.phlippheuer.twitch4j" level="INFO" />
+	<logger name="me.phlippheuer.twitch4j" level="INFO"/>
 
 	<!-- Appender Configuration [Console] -->
 	<appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
@@ -32,9 +32,10 @@
 
 	<!-- Appender Output Configuration -->
 	<root level="debug">
-		<appender-ref ref="CONSOLE" />
+		<appender-ref ref="CONSOLE"/>
 
 		<!-- Uncomment the next line to also log to a file -->
-		<appender-ref ref="FILE"/> -->
+		<appender-ref ref="FILE"/>
+		-->
 	</root>
 </configuration>
